Move QM methods back to correction.py, remove unused files, reimplement DQM to match Cannon fig 2, modify tests accordingly
